% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{idp-package}
\alias{idp}
\alias{idp-package}
\title{idp package}
\description{
Isodat Data Processor (IDP)
}
\details{
User interface to facilitate interacting with isodat data.
}
\examples{
\dontrun{\\code{idp.start()}}
}
\author{
Sebastian Kopf
}
\seealso{
\code{\link{idp.start}}
}

