% Generated by roxygen2 (4.0.2): do not edit by hand
\name{idp.start}
\alias{idp.start}
\title{Start the Isodat Data Processor}
\usage{
idp.start(load = NULL, askReload = TRUE)
}
\arguments{
\item{load}{previous instance of the program OR name of the global variable to look for}

\item{askReload}{whether to ask the user if they want to reload the previous instance
or just reload it without asking}
}
\value{
invisible instance of the idp (can be used for calling things directly)
}
\description{
Start the Isodat Data Processor
}

