% Generated by roxygen2 (4.0.2): do not edit by hand
\name{idp.start_from_script}
\alias{idp.start_from_script}
\title{Start the Isodat Data Processor with an Rscript}
\usage{
idp.start_from_script(...)
}
\description{
Same as idp.start except it makes sure to keep the terminal alive with
a loop until the program actually exits
}

